requirements.txt
python-dotenv
google-cloud-storage
azure-storage-blob
boto3